BREAKING: v2.0
- update Angular 16

BREAKING CHANGE:

In previous versions, `emitInitialResult` emits the result outside of
`NgZone` regardless of `emitInZone` option. This has been changed to be
more consistent that is `emitInZone` option is respected. This is
a small change but it is a breaking change.

If you are using `emitInitialResult: true` and `emitInZone: false`,
things _might not_ work correctly anymore. Please update accordingly